Common spec context
