+++ spec/giraph/extensions/field_spec.rb
@@ -1,44 +1,7 @@
 require 'spec_helper'
 
 describe Giraph::Extensions::Field do
-  class MockResolver; end
-
-  let(:schema) do
-    GraphQL::Schema.new(
-      query: query_root,
-      # mutation: mutation_root
-    )
-  end
-
-  let(:query_root) do
-    GraphQL::ObjectType.define do
-      name 'Query'
-
-      field :test do
-        type types.Int
-        argument :foo, types.String
-        resolve MockResolver
-      end
-    end
-  end
-
-  let(:some_context) do
-    {
-      bar: [1, 2]
-    }
-  end
-
-  let(:query_string) do
-    '{ test(foo: "foo value") }'
-  end
-
-  let(:query_variables) do
-    {}
-  end
-
-  let(:giraph_response) do
-    Giraph::Remote::Response[test: 84]
-  end
+  include_context 'sample graphql structure'
 
   it 'allows assigned resolver to resolve under regular execution' do
     expect(MockResolver).to receive(:call) do |obj, args, ctx|
@@ -52,7 +15,7 @@ describe Giraph::Extensions::Field do
     resp = schema.execute(
       query_string,
       variables: query_variables,
-      context: some_context
+      context: query_context
     )
 
     expect(resp).to eq('data' => { 'test' => 42 })
@@ -64,7 +27,7 @@ describe Giraph::Extensions::Field do
     resp = schema.execute(
       query_string,
       variables: query_variables,
-      context: some_context,
+      context: query_context,
       root_value: giraph_response
     )
